requirements.txt
streamlit
pandas
matplotlib
requests
python-bidi
arabic-reshaper
extra-streamlit-components